make notification compoent

+++ src/pages/Category.tsx
@@ -8,6 +8,8 @@ import {
 import { Table } from "antd";
 import { Edit, Plus, Trash2 } from "lucide-react";
 import { useEffect, useState } from "react";
+import toast from "react-hot-toast";
+import Swal from "sweetalert2";
 
 const Category = () => {
   //! Query
@@ -25,18 +27,11 @@ const Category = () => {
   const [deleteCategory, { isSuccess, error, data }] =
     useDeleteCategoryMutation();
   useEffect(() => {
-    if (isSuccess) {
-      if (data) {
-        alert("Offer Delete Successfully");
-        setOpen(false);
-      }
-    }
-
     if (error) {
       if ("data" in error) {
         const errorData = error as any;
 
-        alert(errorData.data.message);
+        toast.error(errorData.data.message);
       } else {
         console.error("Login error:", error);
       }
@@ -99,12 +94,31 @@ const Category = () => {
   };
 
   const handleDelete = async (id: string) => {
-    try {
-      await deleteCategory(id);
-    } catch (error: any) {
-      console.error(error.message);
+    const confirmation = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    });
+
+    if (confirmation.isConfirmed) {
+      try {
+        await deleteCategory(id);
+        Swal.fire("Deleted!", "Your category has been deleted.", "success");
+      } catch (error: any) {
+        console.error(error.message);
+        Swal.fire(
+          "Error!",
+          "There was an error deleting the category.",
+          "error"
+        );
+      }
     }
   };
+
   return (
     <div>
       <Title>Category Management</Title>
